requirements.txt
pyserial>=3.4
uModbus==1.0.3

